Refactor program so it no longer uses global variable 'mode'
Change setGameMode to setMode, remove setModeVariable and move it's
functionality to updateButton which will now add a class of
'current-mode' to the currently selected button. Update mark function,
and deEmphasizeButton to use getMode, and remove the mode global
variable. Add current-mode class to the Regular Mode button in main so
the program is initialized correctly.